Updated the instruction class to use 32-bit index since 16-bit isn't large enough.
Made printing a slightly prettier for instrumentation code.



git-svn-id: svn+ssh://bono.sdsc.edu/projects/pmac/svnrepos/pmac/trunk/X86ElfInstrumentor@2783 55925d41-9515-0410-9586-c6f35bcc715f

+++ include/Instruction.h
@@ -129,7 +129,7 @@ class Instruction : public Base {
 protected:
     uint8_t instructionType;
     uint8_t source;
-    uint16_t index;
+    uint32_t index;
     char* rawBytes;
     Operand** operands;    
     bool leader;
